clean up tabletserver planbuilder
1. merge insert.go and select.go into dml.go
2. merge plan_type.go into plan.go
3. add some comments and rename variables to follow golint's suggestions
